requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==6.0
django-extensions==3.2.3
django-filter==25.2
django-redis==6.0.0
django-silk==5.4.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gprof2dot==2025.4.14
hiredis==3.3.0
inflection==0.5.1
isort==7.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==10.4.0
PyJWT==2.10.1
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.2.0
